refactor(rust): extract AppState + LockExt into state.rs (slice 2/9)
Move AppState (struct + new/require_key) and the LockExt poison-safe
mutex helper out of lib.rs into state.rs; lib.rs re-exports them so every
crate::AppState / .lock_safe() reference is unchanged. cargo test green (60).
